@@ -1,18 +1,15 @@
 <template>
   <div>
-      Completed Todo: {{ numberOfCompletedTodo }}
+    Completed Todo: {{ numberOfCompletedTodo }}
   </div>
 </template>
 
 <script>
 export default {
-    props: {
-        todos: {
-            type: Array,
-            required: true
-        }    
-    },
     computed: {
+        todos() {
+            return this.$store.state.todos;
+        },
         numberOfCompletedTodo() {
             return this.todos.filter(todo => todo.checked).length;
         }
